refactor: beautify edit and create question, add validation for each question type

+++ src/main/java/de/nordakademie/iaa_multiple_choice/domain/Question.java
@@ -31,7 +31,7 @@ public class Question {
     private Long id;
 
     @Basic
-    private Double points;
+    private Integer points;
 
     @Enumerated(EnumType.STRING)
     private QuestionType type;
